[connector] Rework structure
