+++ internal-enrichment/malbeacon/src/malbeacon_converter.py
@@ -1,0 +1,8 @@
+
+
+class MalbeaconConverter:
+    """
+    Convert data from Malbeacon to STIX 2 object
+    """
+    def __init__(self, helper):
+        self.helper = helper
